Remove fmnist data, move projection to demo, and clean up demo files
